% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scads_func.R
\name{calculate_lambda_local_scATAC}
\alias{calculate_lambda_local_scATAC}
\title{Calculate Local Background (Lambda) for scATAC-seq Peaks (Per-Cell Model)}
\usage{
calculate_lambda_local_scATAC(
  count_matrix,
  peaks,
  gc_content = NULL,
  genome_size = 2.7e+09
)
}
\arguments{
\item{count_matrix}{A sparse matrix (e.g., dgCMatrix) with cells as rows
and peaks as columns.}

\item{peaks}{A \code{GRanges} object containing the coordinates for all peaks.
\strong{CRITICAL: The order of peaks \emph{must} match the
order of columns in \code{count_matrix}.}}

\item{gc_content}{A numeric vector of GC content fraction for each peak,
in the same order as \code{peaks}.}

\item{genome_size}{Numeric, the effective genome size. Default is 2.7e9.}
}
\value{
A \code{data.frame} with peak coordinates and all per-cell
lambda values.
}
\description{
This function calculates a local background accessibility rate (lambda) for
scATAC-seq peaks, similar to the method used in MACS2.
}
\details{
This model calculates all rates on a "per-cell" basis
(i.e., average reads per base pair per cell).
}
