% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastTopics_aux_func.R
\name{compute_lfc_stats2}
\alias{compute_lfc_stats2}
\title{Helper functions for scads}
\usage{
compute_lfc_stats2(
  X,
  F,
  L,
  f0,
  D = matrix(rnorm(1000 * k), 1000, ncol(F)),
  U = matrix(runif(1000 * k), 1000, ncol(F)),
  M = matrix(sample(k, 1000 * k, replace = TRUE), 1000, k) - 1,
  lfc.stat = "le",
  conf.level = 0.68,
  rw = 0.3,
  e = 1e-15,
  verbose = TRUE
)
}
\description{
Helper functions for scads
}
