% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scads_func.R
\name{create_continuous_annotation}
\alias{create_continuous_annotation}
\title{Helper functions for scads}
\usage{
create_continuous_annotation(bim_file, bed_file, annot_file)
}
\description{
Helper functions for scads
}
