% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scads_func.R
\name{de_analysis2}
\alias{de_analysis2}
\title{Helper functions for scads}
\usage{
de_analysis2(
  fit,
  X,
  s = rowSums(X),
  pseudocount = 0.01,
  fit.method = c("scd", "em", "mu", "ccd", "glm"),
  shrink.method = c("ash", "none"),
  lfc.stat = "le",
  control = list(),
  verbose = TRUE,
  ...
)
}
\description{
modified de_analysis function from fastTopics
}
