% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{demo_count_matrix}
\alias{demo_count_matrix}
\title{Demo count matrix}
\format{
A sparse matrix of class \code{dgCMatrix}.
}
\usage{
demo_count_matrix
}
\description{
A sparse peak-by-cell count matrix for chr10, subset to 500 cells.
}
\keyword{datasets}
