% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendre.R
\name{draw_poilog_samples}
\alias{draw_poilog_samples}
\title{Draw samples from poilog}
\usage{
draw_poilog_samples(sad_to_use, scale_mu = 1, scale_sig = 1)
}
\arguments{
\item{sad_to_use}{basis sad}

\item{scale_mu}{factor to scale mu by - to mess with prediction}

\item{scale_sig}{factor to scale sig by}
}
\value{
a sample
}
\description{
Draw samples from poilog
}
