% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendre_utils.R
\name{eucl_rows}
\alias{eucl_rows}
\title{Euclidean distance between vectors}
\usage{
eucl_rows(v1, v2)
}
\arguments{
\item{v1}{row 1}

\item{v2}{row 2}
}
\value{
euclidean distance
}
\description{
Euclidean distance between vectors
}
