% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incorrect_sads.R
\name{generate_even_sad}
\alias{generate_even_sad}
\title{Generate an even SAD}
\usage{
generate_even_sad(sad_to_use)
}
\arguments{
\item{sad_to_use}{sad}
}
\value{
even sad
}
\description{
Generate the most even possible SAD of S species summing to N
}
