% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incorrect_sads.R
\name{generate_exponential_sad}
\alias{generate_exponential_sad}
\title{Generate exponential SAD}
\usage{
generate_exponential_sad(sad_to_use)
}
\arguments{
\item{sad_to_use}{sad}
}
\value{
exponential-like SAD
}
\description{
Generate exponential SAD
}
