% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incorrect_sads.R
\name{generate_precipice_sad}
\alias{generate_precipice_sad}
\title{Generate an SAD like a precipice}
\usage{
generate_precipice_sad(sad_to_use)
}
\arguments{
\item{sad_to_use}{sad}
}
\value{
sad with 1 species with almost all the individuals and every other species with 1 individual
}
\description{
Generate an SAD like a precipice
}
