% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incorrect_sads.R
\name{generate_stepwise_sad}
\alias{generate_stepwise_sad}
\title{Generate stepwise sad}
\usage{
generate_stepwise_sad(sad_to_use)
}
\arguments{
\item{sad_to_use}{sad}
}
\value{
stepwise increasing sad
}
\description{
Generate stepwise sad
}
