% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scads_func.R
\name{get_average_bg}
\alias{get_average_bg}
\title{Estimate “background” accessibility rate from a reference cell‐type peak list}
\usage{
get_average_bg(count_matrix, cell_type_cells, cell_type_peaks)
}
\arguments{
\item{count_matrix}{A peaks-by-cells matrix (dgCMatrix or dense).}

\item{cell_type_cells}{list of cells of the specific cell type}

\item{cell_type_peaks}{Path to a BED file of that cell type’s peaks (three cols: chr, start, end).}
}
\value{
A single numeric: the estimated baseline accessibility probability.
}
\description{
Estimate “background” accessibility rate from a reference cell‐type peak list
}
\examples{
\dontrun{
get_average_bg(cm, "CD8", "cd8_peaks.bed")
}
}
