% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_continuous_annot_beds.R
\name{get_continuous_annot_beds}
\alias{get_continuous_annot_beds}
\title{Generate Continuous BED Files from Topic Annotations}
\usage{
get_continuous_annot_beds(topics_res, output_dir)
}
\arguments{
\item{topics_res}{A list containing at least the factor matrix \code{Fmat}.
\itemize{
\item \code{Fmat}: A matrix containing continuous factor loadings (peaks x topics).
Rows should have names in the format "chr:start-end" (e.g., "chr1:1000-2000").
Columns correspond to topics (e.g., k1, k2, ..., k5).
}}

\item{output_dir}{A character string specifying the base output directory for the BED files.}
}
\value{
A list of output folder paths (\code{bed_dir_list}), one for each topic.
}
\description{
This function generates continuous BED files for each topic based on the provided factor matrix \code{Fmat}
from fastTopics. Each BED file is saved into its own folder, and a list of output folder paths is returned.
}
