% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendre_utils.R
\name{get_leg_matrix}
\alias{get_leg_matrix}
\title{Retrieve Legendre polynomials}
\usage{
get_leg_matrix(n, x)
}
\arguments{
\item{n}{integer will retrieve the first n polynomials}

\item{x}{integer vector will evaluate along x scaled to be from -1:1}
}
\value{
`x` by `n` matrix of n polynomials evaluated at `length(x)` points, without the intercept (0th order polynomial)
}
\description{
Retrieves the first `n` Legendre polynomials, evaluated on `x` scaled from -1 to 1.
}
