% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signif_loci.R
\name{get_signif_loci}
\alias{get_signif_loci}
\title{Get Significant Loci from Summary Statistics and LD Blocks}
\usage{
get_signif_loci(
  sumstats_path,
  ldblocks_path,
  pval_threshold = 5e-08,
  max_loci = 150,
  outfile = NULL
)
}
\arguments{
\item{sumstats_path}{Directory to a data frame containing summary statistics. Must include columns:
\itemize{
\item \code{chr}: Chromosome number.
\item \code{pos}: Base pair position.
\item \code{snp}: SNP identifier.
\item \code{pval}: P-value of association.
}}

\item{ldblocks_path}{Directory to a data frame containing LD blocks. Must include columns:
\itemize{
\item \code{X1}: Chromosome number.
\item \code{X2}: Start position of the LD block.
\item \code{X3}: End position of the LD block.
\item \code{X4}: Locus identifier.
}}

\item{pval_threshold}{A numeric value specifying the p-value threshold for significance. Default is \code{5e-8}.}

\item{max_loci}{An integer specifying the maximum number of loci to return. Default is \code{150}.}

\item{outfile}{An optional character string specifying the output file path. If provided, the significant loci data frame will be written to this file.}
}
\value{
A data frame of significant loci.
}
\description{
This function identifies significant loci based on summary statistics and LD blocks.
}
