% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendre.R
\name{legendre_approx}
\alias{legendre_approx}
\title{Perform Legendre approximation}
\usage{
legendre_approx(vec, nleg)
}
\arguments{
\item{vec}{an SAD}

\item{nleg}{how many polynomials to include}
}
\value{
lm; Legendre approximation
}
\description{
Perform Legendre approximation
}
