% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendre.R
\name{legendre_distance}
\alias{legendre_distance}
\title{Compute Euclidean distance of Legendre coefficients}
\usage{
legendre_distance(empirical = NULL, draw_from = "mete", nleg = 6, s,
  n, scale_mu = NULL, scale_sig = NULL, compare_sims = T)
}
\arguments{
\item{empirical}{optional empirical vector}

\item{draw_from}{"mete" or "fs" or "poilog"}

\item{nleg}{nb of polynomials}

\item{s}{s0}

\item{n}{n0}

\item{scale_mu}{mess with mu}

\item{scale_sig}{mess with sig}
}
\value{
euclidean distance
}
\description{
Compute Euclidean distance of Legendre coefficients
}
