% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendre.R
\name{legendre_generate}
\alias{legendre_generate}
\title{Generate a (scaled) SAD from Legendre coefficients}
\usage{
legendre_generate(coeff_values, nleg, nspp, scale_pos = 2400)
}
\arguments{
\item{coeff_values}{coeffcicient values}

\item{nleg}{how many polynomials to use (for shiny app; otherwise implicit in coeff_values)}

\item{nspp}{how many speices}
}
\value{
vector of scaled abundances
}
\description{
Generate a (scaled) SAD from Legendre coefficients
}
