% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_sads.R
\name{loglik_sad}
\alias{loglik_sad}
\title{logLik of an sad from METE}
\usage{
loglik_sad(focal_sad, mete_sad_dist)
}
\arguments{
\item{focal_sad}{sad vector}

\item{mete_sad_dist}{result of meteR::sad(meteESF)}
}
\value{
loglikelihood
}
\description{
logLik of an sad from METE
}
