% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sads.R
\name{make_sad_samples_long}
\alias{make_sad_samples_long}
\title{Convert sample dfs to long}
\usage{
make_sad_samples_long(sad_samples)
}
\arguments{
\item{sad_samples}{result of sample_FS, sample_METE}
}
\value{
long df
}
\description{
Convert sample dfs to long
}
