% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_list.R
\name{make_samples_list}
\alias{make_samples_list}
\title{Make a list for comparison}
\usage{
make_samples_list(i, first_samples, second_samples, random = FALSE)
}
\arguments{
\item{i}{index}

\item{first_samples}{first set}

\item{second_samples}{second set}

\item{random}{T/F whether to assign pairs randomly}
}
\value{
list of rows as a list
}
\description{
Make a list for comparison
}
