% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sads.R
\name{plot_sad_samples}
\alias{plot_sad_samples}
\title{Plot SAD samples}
\usage{
plot_sad_samples(long_sad_samples, sample_type = NULL)
}
\arguments{
\item{long_sad_samples}{from `make_sad_samples_long`}

\item{sample_type}{for plot title}
}
\value{
the plot
}
\description{
Plot SAD samples
}
