% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_sads.R
\name{r2}
\alias{r2}
\title{Calculate a single r2
From `conditionalsads`}
\usage{
r2(focal_sad, compare_sad)
}
\arguments{
\item{focal_sad}{ascending RAD ("empirical")}

\item{compare_sad}{ascending RAD from samples}
}
\value{
r2
}
\description{
Calculate a single r2
From `conditionalsads`
}
