% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ldsc.R
\name{run_ldsc}
\alias{run_ldsc}
\title{Run LDSC Pipeline}
\usage{
run_ldsc(
  polyfun_path,
  sumstats_path,
  n,
  trait,
  onekg_path,
  bed_dir,
  baseline_dir,
  weights_dir,
  out_dir
)
}
\arguments{
\item{polyfun_path}{A character string specifying the directory to the PolyFun package.}

\item{sumstats_path}{A character string specifying the directory with the cleaned summary statistics file.}

\item{n}{Number of samples in the GWAS.}

\item{trait}{GWAS trait (e.g. SIM for simulated)}

\item{onekg_path}{A character string specifying the directory to 1000G genotypes data}

\item{bed_dir}{A character string specifying the directory for the BED files.}

\item{baseline_dir}{A character string specifying the directory to baseline annotations for LDSC}

\item{weights_dir}{A character string specifying the directory to weights provided for LDSC}

\item{out_dir}{A character string specifying the output directory for LDSC enrichment results.}

\item{rcode_path}{A character string specifying the directory to Rscripts for running LDSC}
}
\value{
None. LDSC results are written to the specified output directory.
}
\description{
This function runs the LDSC pipeline using PolyFun for enrichment analysis.
}
