% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sldsc.R
\name{run_sldsc}
\alias{run_sldsc}
\title{Run S-LDSC Pipeline: Calculate disease cell score using hg19 BED and Baseline v2.2}
\usage{
run_sldsc(
  chrs,
  polyfun_path,
  ldsc_path,
  sumstats_path,
  n,
  trait,
  onekg_path,
  bed_dir,
  baseline_dir,
  frqfile_pref,
  hm3_snps,
  weights_pref,
  out_dir
)
}
\arguments{
\item{chrs}{Number of chromosomes for running S-LDSC (Default is 1:22).}

\item{sumstats_path}{Directory containing the cleaned GWAS summary statistics file (format: \code{<trait>_sumstats.txt.gz}).}

\item{n}{Sample size of GWAS.}

\item{trait}{GWAS trait name (e.g., "SIM", "IBD").}

\item{onekg_path}{Prefix to 1000G plink files in hg19 (e.g., \code{"/LDSCORE/zenodo/1000G_EUR_Phase3_plink/1000G.EUR.QC."}).}

\item{bed_dir}{Directory containing the user-supplied BED file(s) in hg19. Only the first \code{.bed} found is used.}

\item{baseline_dir}{Prefix to 1000G baseline v2.2 annotation in hg19 (e.g., \code{"/LDSCORE/zenodo/1000G_Phase3_baselineLD_v2.2_ldscores/baselineLD."}).}

\item{frqfile_pref}{Prefix to the 1000G .frq or .afreq files in hg19 (e.g., \code{"/LDSCORE/zenodo/1000G_Phase3_frq/1000G.EUR.QC."}).}

\item{hm3_snps}{Path to the HapMap3 no-MHC SNP list in hg19 (e.g., \code{"/LDSCORE/zenodo/hm3_no_MHC.list.txt"}).}

\item{weights_pref}{Path to the 1000G weights in hg19 (e.g., \code{"/LDSCORE/zenodo/1000G_Phase3_weights_hm3_no_MHC/weights.hm3_noMHC."}).}

\item{out_dir}{Directory where outputs are saved. The function creates \code{annotations/<trait>} and \code{results} subdirectories.

#' @details
\strong{Step-by-step}:
\enumerate{
\item Munge summary stats \code{<trait>_sumstats.txt.gz} into a parquet with
\code{munge_polyfun_sumstats.py}.
\item For each chromosome \code{1..22}:
\itemize{
\item Create annotation (\code{make_annot.py}) from the user .bed plus the .bim for that chromosome.
\item Compute LD scores (\code{ldsc.py --l2}) for the annotation.
}
\item Finally, run \code{ldsc.py --h2} referencing both the newly created annotation
(\code{out_dir/annotations/<trait>/<trait>.<chr>}) and \code{baseline_dir},
providing \code{frqfile_pref} for the allele frequency files and the \code{hm3_snps}
for SNP filtering.
}}

\item{polyfun_code_dir}{Directory to access polyFUN code scripts (e.g., \code{munge_polyfun_sumstats.py}); must be python3-compatible.}

\item{ldsc_code_dir}{Directory to access LDSC code scripts (e.g., \code{make_annot.py}, \code{ldsc.py}).}
}
\value{
No object is returned. All intermediate files (\code{.annot.gz}, \code{.ldscore.gz}) and final
LDSC results (\code{.log}, \code{.results}) are written to \code{out_dir}.
}
\description{
This function runs the S-LDSC pipeline using PolyFun scripts for:
\enumerate{
\item Munging summary statistics ("munge_polyfun_sumstats.py"),
\item Generating annotation files from a user-supplied BED (hg38) via "make_annot.py",
\item Computing LD scores with "ldsc.py --l2",
\item Performing final heritability estimation ("ldsc.py --h2")
by combining the custom annotation with baseline v2.2 annotations.
}
}
\examples{
\dontrun{
run_sldsc(chrs = 1:22,
          polyfun_code_dir = "/path/polyfun/",
          ldsc_code_dir = "/path/ldsc/",
          sumstats_path = "/path/sumstats/",
          n = 60000, trait = "SIM",
          onekg_path = "/LDSCORE/zenodo/1000G_EUR_Phase3_plink/1000G.EUR.QC.",
          bed_dir = "/path/my_bedfiles/",
          baseline_dir = "/LDSCORE/zenodo/1000G_Phase3_baselineLD_v2.2_ldscores/baselineLD.",
          frqfile_pref = "/LDSCORE/zenodo/1000G_Phase3_frq/1000G.EUR.QC.",
          hm3_snps = "/LDSCORE/zenodo/hm3_no_MHC.list.txt",
          weights_pref = "/LDSCORE/zenodo/1000G_Phase3_weights_hm3_no_MHC/weights.hm3_noMHC.",
          out_dir = "results/")
}


}
