% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-fs.R
\name{sample_feasibleset}
\alias{sample_feasibleset}
\title{Draw from feasible set}
\usage{
sample_feasibleset(s, n, nsamples, distinct = TRUE)
}
\arguments{
\item{s}{how many species}

\item{n}{how many individuals}

\item{nsamples}{how many samples to draw}

\item{distinct}{TRUE/FALSE filter to distinct vectors. If true may not return nsamples samples.}
}
\description{
Use `feasiblesads` to generate samples from the feasible set. Building on code from `diazrenata/conditionalsads`
}
