% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-mete.R
\name{sample_METE}
\alias{sample_METE}
\title{Draw from METE prediction}
\usage{
sample_METE(s, n, nsamples, distinct = TRUE)
}
\arguments{
\item{s}{how many species}

\item{n}{how many individuals}

\item{nsamples}{how many samples to draw}

\item{distinct}{TRUE/FALSE filter to distinct vectors. If true may not return nsamples samples.}
}
\value{
matrix of samples. rows are samples, columns are species
}
\description{
Use `meteR` to generate the SAD from METE and draw samples from it
}
