% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendre_utils.R
\name{scale_vector}
\alias{scale_vector}
\title{Scale a vector for Legendre approximation}
\usage{
scale_vector(vec)
}
\arguments{
\item{vec}{Vector (an SAD) to scale}
}
\value{
dataframe of `x` = `(-1:1, length = length(vec))` and `val` = `vec / sum(vec)`
}
\description{
Scale a vector for Legendre approximation
}
