% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_python.R
\name{setup_python_env}
\alias{setup_python_env}
\title{Set up Python Environment for scads Package}
\usage{
setup_python_env(python_path = NULL)
}
\description{
This function creates and configures a Python environment with the required packages.
}
