% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendre_utils.R
\name{ssqe}
\alias{ssqe}
\title{Calculate sum squared errorfor two vectors}
\usage{
ssqe(a, b, scale = T)
}
\arguments{
\item{a}{vector}

\item{b}{vector of `length(b) == length(a)`}
}
\value{
ssq of a and b
}
\description{
Calculate sum squared errorfor two vectors
}
