% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cs.R
\name{get_cs}
\alias{get_cs}
\title{Compute Single-Cell Level Scores (c_i)}
\usage{
get_cs(topic_res, ldsc_res_dir, trait, nTopics)
}
\arguments{
\item{topic_res}{Results from run_fastTopics, with Pmat (J×K) and Lmat (I×K)}

\item{ldsc_res_dir}{Directory containing k*_output/results/Trait.results}

\item{trait}{Trait name used in LDSC results files.}

\item{nTopics}{Number of topics (K).}
}
\value{
A list with
\itemize{
\item cs: vector of length I (cells)
\item var_cs: variance of each cs\link{i}
\item ldsc_res_table: the raw tau_display table from LDSC
}
}
\description{
Compute Single-Cell Level Scores (c_i)
}
