% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cs.R
\name{get_cs}
\alias{get_cs}
\title{Compute Single-Cell Level Scores (c_i)}
\usage{
get_cs(topic_res, ldsc_res_dir, trait, nTopics)
}
\arguments{
\item{topic_res}{Results from run_fastTopics_res, with Pmat (J×K) and Lmat (I×K)}

\item{ldsc_res_dir}{Directory containing k*_output/results/Trait.results}

\item{trait}{Trait name used in LDSC results files.}

\item{nTopics}{Number of topics (K).}
}
\value{
A list with
\itemize{
\item cs: cell scores (vector of length I)
\item z_cell: z-score accounting for variance
\item p_cell: p-value for assessing score reliability
\item ldsc_res_table: the raw tau_display table from S-LDSC
\item ash_res: results from running ASH on S-LDSC enrichment estimates
\item cs_dat: intermediate estimates for calculating cell score
}
}
\description{
Compute Single-Cell Level Scores (c_i)
}
