% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cs.R
\name{get_cs}
\alias{get_cs}
\title{Compute Single-Cell Level Scores (c_i)}
\usage{
get_cs(p_jk, l_ik, tau_Ck)
}
\arguments{
\item{p_jk}{Matrix of probabilities of peaks in topics (peaks x topics), obtained from the F matrix of fastTopics.}

\item{l_ik}{Matrix of topic proportions for each cell (cells x topics), obtained from the L matrix of fastTopics.}

\item{tau_Ck}{Vector of per-topic heritability contributions (\eqn{\tau_{C_k}}) of length K (number of topics).}
}
\value{
A vector of single-cell level scores (\eqn{c_i}) of length I (number of cells).
}
\description{
This function computes the single-cell level scores (\eqn{c_i}) based on the fastTopics model outputs
and other required parameters.
}
\details{
The function computes the scores based on the following corrected approximation:
\deqn{
c_i = \frac{ \left( \frac{E(M_i)}{E(N_i)} \left[ 1 - \frac{\text{Cov}(M_i, N_i)}{E(M_i) E(N_i)} + \frac{\text{Var}(N_i)}{[E(N_i)]^2} \right] \right) }{ \left( \frac{ \sum_{j=1}^{J} h_j }{ J } \right) }
}
where:
\itemize{
\item \eqn{E(M_i)} and \eqn{E(N_i)} are the expected values.
\item \eqn{\text{Cov}(M_i, N_i)} is the covariance between \eqn{M_i} and \eqn{N_i}.
\item \eqn{\text{Var}(N_i)} is the variance of \eqn{N_i}.
}
}
