% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ldsc.R
\name{run_ldsc}
\alias{run_ldsc}
\title{Run LDSC Pipeline}
\usage{
run_ldsc(
  polyfun_path,
  sumstats_path,
  n,
  mss_output,
  ldblocks_path,
  sig_loci_output = NULL
)
}
\arguments{
\item{polyfun_path}{Path to the PolyFun package.}

\item{sumstats_path}{Path to the cleaned summary statistics file.}

\item{n}{Number of samples in the GWAS.}

\item{mss_output}{Munged sumstats file.}

\item{ldblocks_path}{Path to file with LD blocks.}

\item{sig_loci_output}{Signficant loci output file.}
}
\value{
enrichment
}
\description{
This function runs the LDSC pipeline using PolyFun for enrichment analysis.
}
