% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scads.R
\name{scads}
\alias{scads}
\title{scads: A function for using scATAC-seq data and GWAS sumstats to calculate disease cell score}
\usage{
scads(
  count_matrix,
  nTopics = 10,
  bed_dir,
  sumstats_dir,
  gwas_nsamps,
  gwas_trait,
  outdir,
  ...
)
}
\arguments{
\item{count_matrix}{A peak-by-cell count matrix.}

\item{nTopics}{The number of topics (clusters) to fit. Default is 10.}

\item{bed_dir}{}

\item{...}{Additional arguments passed to internal functions.}
}
\value{
A list containing the fitted topic model, enrichment results, and cell scores.
}
\description{
scads: A function for using scATAC-seq data and GWAS sumstats to calculate disease cell score
}
\examples{
result <- scads(count_matrix, GWAS_ss, LD_reference, nTopics = 15)

}
